In some networks like alastria, networkid can be bigger than integer (https://github.com/alastria/alastria-node/wiki/private_transactions_in_alastria). Change necessary from integer to Long.

+++ cakeshop-api/src/main/java/com/jpmorgan/cakeshop/controller/NodeController.java
@@ -108,7 +108,7 @@ public class NodeController extends BaseController {
             }
 
             if (!StringUtils.isEmpty(jsonRequest.getNetworkId())) {
-                nodeSettings.networkId(Integer.parseInt(jsonRequest.getNetworkId()));
+                nodeSettings.networkId(Long.parseLong(jsonRequest.getNetworkId()));
             }
 
             if (jsonRequest.getCommittingTransactions() != null) {
